Refactor imports for @hedvig-ui/Button, again
