+++ shared/hedvig-ui/Button/button.stories.tsx
@@ -1,7 +1,7 @@
 import { Button, buttonColorMap, ButtonProps } from '@hedvig-ui'
 import { action } from '@storybook/addon-actions'
 import { boolean, select } from '@storybook/addon-knobs'
-import { lightTheme } from 'hedvig-ui/themes'
+import { lightTheme } from 'hedvig-ui'
 import React from 'react'
 import { WithStory } from '../story-utils'
