+++ src/components/member/tabs/campaigns-tab/referrals/ReferralsInfo.tsx
@@ -1,11 +1,10 @@
-import { Capitalized, Card, CardsWrapper, Placeholder } from '@hedvig-ui'
+import { Badge, Capitalized, Card, CardsWrapper, Placeholder } from '@hedvig-ui'
 import { MemberReferral, ReferralInformation } from 'api/generated/graphql'
 import {
   BadgeRow,
   SmallTopSpacing,
   TableHeadline,
 } from 'components/member/tabs/campaigns-tab/styles'
-import { Badge } from 'hedvig-ui/Badge/badge'
 import React from 'react'
 import { Link } from 'react-router-dom'
 import { MembersReferredTable } from './MembersReferredTable'
