+++ src/types/emotion.d.ts
@@ -1,5 +1,5 @@
 import '@emotion/react'
-import { lightTheme } from 'hedvig-ui/themes'
+import { lightTheme } from '@hedvig-ui'
 
 declare module '@emotion/react' {
   type ThemeRecord = Record<keyof typeof lightTheme, string>
